+++ packages/builder/builder-webpack-provider/src/types/config/output.ts
@@ -9,6 +9,7 @@ export type DistPathConfig = {
   html?: string;
   image?: string;
   media?: string;
+  server?: string;
 };
 
 export type FilenameConfig = {
@@ -89,15 +90,19 @@ export type RemOptions = Partial<{
 
 export type ExternalsOptions = WebpackConfig['externals'];
 
+export type Charset = 'ascii' | 'utf8';
+export type SvgDefaultExport = 'component' | 'url';
+export type LegalComments = 'none' | 'inline' | 'linked';
+
 export interface OutputConfig {
   copy?: CopyPluginOptions | CopyPluginOptions['patterns'];
   distPath?: DistPathConfig;
   filename?: FilenameConfig;
-  charset?: 'ascii' | 'utf8';
+  charset?: Charset;
   polyfill?: Polyfill;
   assetPrefix?: string;
   dataUriLimit?: number | DataUriLimit;
-  legalComments?: 'none' | 'inline' | 'linked';
+  legalComments?: LegalComments;
   cleanDistPath?: boolean;
   disableMinimize?: boolean;
   disableSourceMap?: boolean;
@@ -110,7 +115,7 @@ export interface OutputConfig {
   enableInlineScripts?: boolean;
   enableInlineStyles?: boolean;
   overrideBrowserslist?: string[];
-  svgDefaultExport?: 'component' | 'url';
+  svgDefaultExport?: SvgDefaultExport;
   assetsRetry?: AssetsRetryOptions;
   convertToRem?: boolean | RemOptions;
   externals?: ExternalsOptions;
@@ -132,5 +137,5 @@ export interface NormalizedOutputConfig extends OutputConfig {
   enableCssModuleTSDeclaration: boolean;
   enableInlineScripts: boolean;
   enableInlineStyles: boolean;
-  svgDefaultExport: 'component' | 'url';
+  svgDefaultExport: SvgDefaultExport;
 }