+++ packages/builder/builder-webpack-provider/src/types/config/tools.ts
@@ -103,7 +103,7 @@ export interface ToolsConfig {
   babel?: ToolsBabelConfig;
   terser?: ToolsTerserConfig;
   tsLoader?: ToolsTSLoaderConfig;
-  tsChecker?: false | ToolsTSCheckerConfig;
+  tsChecker?: boolean | ToolsTSCheckerConfig;
   devServer?: ToolsDevServerConfig;
   minifyCss?: ToolsMinifyCssConfig;
   htmlPlugin?: false | ToolsHtmlPluginConfig;