+++ packages/builder/builder-webpack-provider/src/webpackPlugins/ProgressPlugin/helpers/type.ts
@@ -8,6 +8,8 @@ export type Props = {
   char: string;
   width: number;
   buildIcon: string;
+  errorIcon: string;
+  errorInfo: string;
   finishIcon: string;
   finishInfo: string;
   message: string;
@@ -17,6 +19,8 @@ export type Props = {
   messageColor: typeof Color;
   id: string;
   maxIdLen: number;
+  compileTime: string | null;
+  hasErrors: boolean;
 };
 
 export type BusOption = {