+++ packages/builder/builder-webpack-provider/tests/plugins/tsChecker.test.ts
@@ -1,31 +1,88 @@
 import { expect, describe, it } from 'vitest';
-import { PluginTsLoader } from '../../src/plugins/tsLoader';
+import { PluginTsChecker } from '../../src/plugins/tsChecker';
 import { createStubBuilder } from '../../src/stub';
+import type { Context } from '../../src/types';
 
 describe('plugins/tsChecker', () => {
-  it("should't set ts-checker", async () => {
+  const context = {
+    tsconfigPath: '/path/tsconfig.json',
+  } as Context;
+
+  it('should disable ts-checker when tools.tsChecker is false', async () => {
     const builder = await createStubBuilder({
-      plugins: [PluginTsLoader()],
+      plugins: [PluginTsChecker()],
+      context,
       builderConfig: {
         tools: {
           tsChecker: false,
         },
       },
     });
-    const config = await builder.unwrapWebpackConfig();
 
-    expect(config).toMatchSnapshot();
+    expect(
+      await builder.matchWebpackPlugin('ForkTsCheckerWebpackPlugin'),
+    ).toBeFalsy();
   });
 
-  it('should set tsChecker plugin', async () => {
+  it('should enable tsChecker plugin by default', async () => {
     const builder = await createStubBuilder({
-      plugins: [PluginTsLoader()],
+      plugins: [PluginTsChecker()],
+      context,
       builderConfig: {
         tools: {},
       },
     });
-    const config = await builder.unwrapWebpackConfig();
 
+    expect(
+      await builder.matchWebpackPlugin('ForkTsCheckerWebpackPlugin'),
+    ).toBeTruthy();
+  });
+
+  it('should enable tsChecker plugin when tools.tsChecker is true', async () => {
+    const builder = await createStubBuilder({
+      plugins: [PluginTsChecker()],
+      context,
+      builderConfig: {
+        tools: {
+          tsChecker: true,
+        },
+      },
+    });
+
+    expect(
+      await builder.matchWebpackPlugin('ForkTsCheckerWebpackPlugin'),
+    ).toBeTruthy();
+  });
+
+  it('should enable tsChecker plugin when tools.tsChecker is empty object', async () => {
+    const builder = await createStubBuilder({
+      plugins: [PluginTsChecker()],
+      context,
+      builderConfig: {
+        tools: {
+          tsChecker: {},
+        },
+      },
+    });
+
+    expect(
+      await builder.matchWebpackPlugin('ForkTsCheckerWebpackPlugin'),
+    ).toBeTruthy();
+  });
+
+  it('should allow to modify the config of tsChecker plugin', async () => {
+    const builder = await createStubBuilder({
+      plugins: [PluginTsChecker()],
+      context,
+      builderConfig: {
+        tools: {
+          tsChecker: {
+            async: false,
+          },
+        },
+      },
+    });
+    const config = await builder.unwrapWebpackConfig();
     expect(config).toMatchSnapshot();
   });
 });