+++ packages/server/plugin-koa/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @modern-js/plugin-koa
 
+## 1.20.1
+
+### Patch Changes
+
+- Updated dependencies [49515c5]
+  - @modern-js/utils@1.20.1
+  - @modern-js/bff-core@1.20.1
+  - @modern-js/bff-runtime@1.20.1
+
 ## 1.20.0
 
 ### Patch Changes